agents/strands-mcp-assistant/requirements.txt
strands-agents
strands-agents-tools
strands-agents[anthropic]
streamlit